Fix conflicts after merging with master

+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/BIRPackageSymbolEnter.java
@@ -1392,7 +1392,7 @@ public class BIRPackageSymbolEnter {
             readTypeFromCp();
         }
 
-        private BUnionType readUnionType(int flags, int cpI) throws IOException {
+        private BUnionType readUnionType(long flags, int cpI) throws IOException {
             boolean isCyclic = inputStream.readBoolean();
             Name unionNameValue;
             if (isCyclic) {
